src/requirements.txt
zenoml>=0.6.3
pydantic<2.0.0
pyarrow
gradio>=4.0.0
transformers>=4.35.0